Refactored operators.
ProdUpdate is now just an element wise multiplication operator.
ProdUpdates at the python level are turned into an elementwise
multiplication followed by a dotinc operation.

DotInc comes in two varieties: DotIncMV, DotIncVV.
The former is used only in the case where one of the two signals
absolutely has to be treated as a matrix.

This is just a more elegant way to use the UBlas functions to
implement the functionality of the python operators.