String-based validation of MPI sends.
Simulator chunks can be converted into strings which should be the
same regardless of the node they're instantiated on. In particular,
they do not include pointers/addresses.

After we send a chunk to a worker, that worker sends back the string
generated by the chunk it receives, and we comparate that with the
string generated by the original chunk on the master node, make
sure they're the same.

Added MpiSimulatorChunk::to_string.
Added << overload for MpiSimulatorChunk.

Added << overload for Probe<T>. (Slightly tricky adding a friend to
a class template).