Improved handling of PyFunc operators.
Can now handle functions that always return None (still calls the
function, but no longer looks for a signal to set).

Also fails properly if the function outputs something valid on the first
evaluation but None later on.

PyFunc genrally handled more robustly.

Also, finding the C++ key for an object is now done by a function
make_key so that it can easily be changed. Fixed a bug related to this,
where it should get the key by looking it that base signal, and it
wasn't always doing that.