nengo_mpi now passes operator tests.
These test some of the operators that are at the core of nengo_mpi.
These broke after making the changes required to work with updated
version of nengo, and now they're fixed again.

Created new type of Simulator, TestSimulator, which allows individual
operators to be added, rather than requiring that a Nengo network be
passed in. Saves us from having to handle this case in the real
nengo_mpi.Simulator.