Should be able to build arbitrarily large models.
A big change to make it so that signals that are created by the builder
are ``shipped out'' to (the remote MPI process they will live on during
the simulation) as soon as they are created. This means we never have
the entire model residing on a single machine at any given time, which
means we are not bound by the memory of any single machine.

Also changed how operators are sent to remote machines...they are now
sent using strings summarizing their entire contents. These strings are
constructed at the python level. We will no longer need the C++
operator serialization functions, so those should be gotten rid of in a
future commit. This change was made to make the incremental building of
the models easier.