Handle SignalViews properly.
A large commit to make nengo_mpi handle SignalViews properly. Previously
it was essentially ignoring them.

Now, instead of just storing data in a Matrix, we have two data
structures. The first is a BaseMatrix, which actually stores the data.
These are analogous to Signals in the refimpl. The second is a Matrix,
which is effectively a view into a BaseMatrix, and so these are
analogous to SignalViews. All operators now do their computation in
terms of the Matrix data structure, but the data is originally added to
the nengo_mpi chunks as a BaseMatrix, using the add_base_signal function.

Now, when we build operators and supply the strings referencing signals,
we have to provide additional information to completely specify a
SignalView. In particular, we add shape, stride and offset information.

I've also taken out all the serialization code since it is no longer
used at all.

I've also significantly reduced the use of pointers throughout the code,
which has cleaned things up quite a bit, and should not have any
significant performance penalty.