Added support for running simulations w/o python.
Added the file stand_alone.cpp, a simple script to create and run a
simulator without using python. The data for the network it creates is
loaded from a file. Said file must be created by running the library
with python in the usual way, taking care to supply a filename to the
Simulator constructor call. The idea is that this will make some things,
easier, especially hooking up to a debugger.