Added --mca pml ob1 to mpirun call in qsubmit.py.
Was previously having trouble getting the code to run with more
than 128 processors...it would hang during the MPI_COMM_SPAWN call.
Contacted scinet support, found out that when using more than 128
processors, it automatically switches to a different protocal for
point-to-point communication (pml: Point-to-point management layer).
Problem is, the original process that is calling MPI_COMM_SPAWN is still
using the original protocol, and then clearly bad things are going to
happen. Adding this command line argument forces all processors to use
the same protocol. ob1 can be replaced by cm...haven't tested the effect
of doing this though.