Remove dependence on boost MPI.
Now just using the C interface to MPI, which is better anyway since
boost mpi does not support newer features of MPI (e.g. see
http://lists.mpi-forum.org/mpi-forum/2011/12/1396.php).

In the meantime, fixed a problem that I think was slowing simulations
down severely. Did some profiling with Allinea MAP, and found that the
code was spending much of its time in mpi allocate and deallocate calls.
I think this may have been because in the MPI sends and receives in the
MPISend and MPIRecv operators, we were sending boost objects...this
forced us to use boost serialization, and I think the communication
ended up not being done efficiently.