Altered build funcs to deal with node connections.
nengo_mpi is currently allergic to pyfuncs: its best to use them as
little as possible. The refimpl nengo builder will use a pyfunc to
implement certain slices (in the process of building a connection)
when the pre-object of the connection is a Node. Altered nengo_mpi's
version of build_node to replace such connections with connections
that are functionally equivalent, but will not use a pyfunc, by just
expanding the transformation matrix.

Also, altered nengo_mpi's build_ensemble function to disallow ensembles
in direct-mode.