Added HDF5 output capabilities.
SimulatorLog class for use in the serial case (one component), and
ParallelSimulatorLog class for use in parallel case. Can pass in a
filename to write the output to, and if no filename is specified then
nothing is written.

Some fairly large changes to make this possible. MpiInterface is now
MpiSimulator which inherits from Simulator. Regular Simulator MUST be
used in the serial case.