Added MpiBuilder class.
Now we plug in the custom operators to this MpiBuilder class, instead of
into the reference implementation builder. This makes it OK to import
nengo_mpi in the same python script where you run a reference
implementation, which was not the case before.

Also, minor change, added a default_partitioner property to the
Partitioner class.