Cleanup after deadline sprint.
Made some somewhat rushed changes trying to make the April 24th
deadline, this commit smooths out some of those rough edges, but
preserves most of the added functionality.

Chunks are now completely responsible for creating their own simulation
logs. We're usually loading from a file, so the chunks will have access
to the probe data, and so there's no problem. The exception, for now, is
when running through python.

Significant bugs:

For ParalellSimulationLog, the workers were calling the constructor
meant for the master, so the `processor` field of the ParalellSimulationLog
was not being set properly (it ended up being 0 for all processes). So
merged both constructors into one, and add `processor` as an argument
for it.

Note: this commit breaks running MPI simulations through python, but
fortunately we're not using that aspect much right now. That aspect
needs some extra attention anyway, now that we're using HDF5 file IO
to send simulation objects to other processors instead of explicitly
sending messages. Will probably have to add some code to have the master
process write an HDF5 file first, and then have all other processes read
it.