Finished fixing running from python.
All operators are now equipped with an index, a floating point number
giving their order of execution. This is determined in python using
the topological sort. During a finalization step, we sort the
operators by their index to get the correct execution order; this
prevents us from having to take care to add the operators to the mpi
simulator in the correct order.

This helped adding PyFuncs; now that we are always loading the network
from a file, we have to add the PyFuncs after the network has been
loaded, which means that they will be out of order; we needed some
mechanism to make sure we could put them in the correct order
afterwards.