First pass at separating shared libs.
Want to make it possible to build a shared library that runs in serial,
even if mpi is not installed on the system. This is the first pass at
separating the non-mpi parts from the mpi parts enough to make this a
reality. Primary difficulty is in separating our the chunk class,
everything else was already pretty good.