Fixed testing bug.
When importing tests from nengo, was getting a KeyError: 'compare'. In
nengo/tests/conftest.py we can see multiple checks for a 'compare' key.
This key is created by the code in nengo/tests/options.py, but that code
is apparently not executed when importing tests. So I just copy-pasted
into nengo_mpi's conftest.py. One other small change had to be made, as
the 'compare' key was still not present. Had to add a default=None to
the compare argument in the code stolen from options.py.