Changed format of HFD5 files.
Before this commit, we were storing each signal in its own HDF5 data
set. This seemed to slow things down quite a bit on Bluegene, so now
all the signals for a processor are stored in one big data set. Also
required the creation of supplementary data sets storing signals sizes,
signals keys and signal labels, all of which were previously stored as
attributes of the signal data set.