Improved how strings are handled in hdf5 files.
Was previously using 2-D arrays of characters, each row representing
a string but since HDF5 can't handle variable length data types when
running in parallel, the width of those arrays had to be as long as
the biggest string. Got rid of this scheme, replaced it with a scheme
where we use 1-D arrays of characters with null terminators marking the
boundaries between successive strings.