Fixed bug in nengo_mpi/__main__.py.
Previously, the called scripts were not using the true global namespace
(i.e. what you get when you do "import __main__; __main__.__dict__")
as their global namespace (i.e. what is passed in as globals to the exec
call). Consequently, the called scripts were not populating the true global
namespace when defining things, which was causing anything that made use
of the "import __main__" pattern to break (e.g. pickling requires that
classes to be pickled are defined in __main__).