Turned mpi_sim.so into a python C++ extension.
Previously it was technically just a shared library, and could not be
directly imported as a python module. We had to use ctypes, which is
really only meant for wrapping system libraries and required
manipulation of LD_LIBRARY_PATH in order for python to be able to find
it at runtime.